// src/components/article/ImageSlider/types.ts
export interface PaginationData {
    total: number;
    page: number;
    pageSize: number;
    totalPages: number;
}

export interface ThemeStyles {
    bg: string;
    text: string;
    tabBg: string;
    tabBorder: string;
    iconColor: string;
    hoverColor: string;
    activeColor: string;
    imageBg: string;
    buttonBg: string;
    buttonHover: string;
    counterBg: string;
    thumbBg: string;
    thumbBorder: string;
    thumbActive: string;
    thumbHover: string;
    scrollbar: string;
    gradientFrom: string;
    scrollButtonBg: string;
    scrollButtonHover: string;
}